[1824] Fix issue with buying Staatsbahn in SR1, fixed #12076
There was a bug so that if buying SD share in SR1, you became president.
This caused issues later, when SD was formed, as president share was
no longer among the reserved.

This bug was fixed earlier for SR2+, but was missing for SR1.
Have now refactored a bit so that methods shared between SR1 and SR2+
can be put in the BuySellParShares base class of 1824. This means it
is also available for 1824 Cisleithania as the Buy Sell steps inherits
from 1824.
