+++ lib/engine/game/g_1824/step/buy_sell_par_exchange_shares.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-require_relative '../../../step/buy_sell_par_shares'
+require_relative 'buy_sell_par_shares'
 
 module Engine
   module Game
     module G1824
       module Step
-        class BuySellParExchangeShares < Engine::Step::BuySellParShares
+        class BuySellParExchangeShares < G1824::Step::BuySellParShares
           EXCHANGE_ACTIONS = %w[buy_shares].freeze
           BUY_ACTION = %w[special_buy].freeze
           PURCHASE_ACTIONS = Engine::Step::BuySellParShares::PURCHASE_ACTIONS + [Action::SpecialBuy]
@@ -38,10 +38,6 @@ module Engine
             return EXCHANGE_ACTIONS if @game.mountain_railway?(entity) && @game.mountain_railway_exchangable?
 
             []
-          end
-
-          def visible_corporations
-            @game.sorted_corporations.reject { |c| c.type == :minor || c.type == :construction_railway }
           end
 
           def can_buy?(entity, bundle)
@@ -141,14 +137,6 @@ module Engine
             @game.payoff_player_loan(player, payoff_amount: action.amount)
             @round.last_to_act = player
             @round.current_actions << action
-          end
-
-          def allow_president_change?(corporation)
-            # In case of Staatsbahn, president change is only allowed after formation
-            return false if @game.staatsbahn?(corporation) && !corporation.floated?
-
-            reserved = corporation.reserved_shares
-            reserved.none? { |s| s.percent == 20 }
           end
 
           private
